Revert "take out the removal of the postfix_mailvolume munin plugin. I do not see what the bug is, and this interferes with local installation of the plugin (causes duplicate definitions)"
It looks like the reason this is here is because on a default debian
install, without postfix installed, this postfix munin module is
activated, leading to empty graphs. So rather that leave non-postfix
people in this situation, we can instead leave this in and the class
can be overridden locally.

This reverts commit ebfde169e0ef4fc749cd6804e707103fab7dbf78.

Conflicts:

	manifests/client/debian.pp