Fix requirements for plugins
When including munin::host without explicitly including munin::client,
puppet fails to run with the following message:

  Could not find dependent Service[munin-node] for
    File[/etc/munin/plugins] at
    /tmp/vagrant-puppet/modules-0/munin/manifests/plugins/setup.pp:14

This is because munin::host declares a munin plugin, and in turn this
one creates a dependency on the 'munin-node' package and service. Since
this dependency is not fulfilled, the run fails.

The munin::plugins::setup class needs to fulfill its own dependencies by
including munin::client.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>